Micro-optimisation: increase laziness.
Turns out, this saves another roughly 5%. It's a bit unclear to me why, because
eventually all of it gets evaluated, but maybe the timing is relevant.

This probably doesn't actually help much, but it simplifies the code so
you don't need to think about why we're forcing evaluation at this
point. Benchmarking showed some improvement, but not significant enough or
stable enough to matter. This is more of a cleanup.