Heavily constrain the GLSL syntax accepted.
We now emit and parse something very close to SSA. The goal is to reduce
the number of bits required to describe a statement to the minimum. This
means for example that there is no recursion at all in the AST for
expressions. Statements do have recursion in `if`, and this is required.
We sometimes even emit `if` within `if`, so the recursion depth isn't
limited. This is fine, because we can treat the `if` body almost like
a function.